Moved utils into utilities file
