+++ utilities/utilities.py
@@ -6,7 +6,6 @@ import json
 import os
 
 from time import sleep
-
 # import genie either from the local project in pycharm or from virtual env
 try:
     from source import genie_api_setup
@@ -163,3 +162,93 @@ def _wait_for_and_assert_dae_simulation_mode(mode):
     else:
         if g.get_dae_simulation_mode() != mode:
             raise AssertionError("Could not set DAE simulation mode to {}".format(mode))
+
+
+def as_seconds(time):
+    """
+    Convert a up time to seconds
+    Args:
+        time: in format HH:MM:SS
+
+    Returns: time in seconds
+
+    """
+    bits = time.split(":")
+    seconds = 0
+    for bit in bits:
+        seconds *= 60
+        seconds += int(bit)
+
+    return seconds
+
+
+def _start_stop_ioc_is_a_start(is_a_start, ioc_name):
+    """
+    Start or stop and ioc dependent on whether it "is_a_start"
+    Args:
+        is_a_start: True start the ioc; False stop the ioc
+        ioc_name: name of the ioc
+
+    Raises:
+        IOError error if IOC does not start/stop after 30 seconds
+
+    """
+    if is_ioc_up(ioc_name) != is_a_start:
+        g.set_pv("CS:PS:{}:{}".format(ioc_name, "START" if is_a_start else "STOP"), 1, is_local=True)
+
+    wait_for_ioc_start_stop(timeout=30, is_start=is_a_start, ioc_name=ioc_name)
+
+
+def start_ioc(ioc_name):
+    """
+    Start the ioc
+    Args:
+        ioc_name: name of the ioc to start
+
+    Raises:
+        IOError error if IOC does not start after 30 seconds
+    """
+    _start_stop_ioc_is_a_start(True, ioc_name)
+
+
+def stop_ioc(ioc_name):
+    """
+    Stop the ioc
+    Args:
+        ioc_name: name of the ioc to stop
+
+    Raises:
+        IOError error if IOC does not stop after 30 seconds
+    """
+    _start_stop_ioc_is_a_start(False, ioc_name)
+
+
+def wait_for_ioc_start_stop(timeout, is_start, ioc_name):
+    """
+    Wait for an ioc to start or stop, if timeout raise a timeout error
+    Args:
+        timeout: time to wait before there is a timeout error
+        is_start: True to wait for ioc to be started; False to wait for stop
+        ioc_name: name of the ioc to wait for
+
+    Raises:
+        IOError error if IOC does not start/stop after timeout
+    """
+    for count in range(timeout):
+        g.waitfor_time(seconds=1)
+        print("Waited {}s for IOC to {}".format(count, "start" if is_start else "stop"))
+        if is_ioc_up(ioc_name) == is_start:
+            break
+    else:
+        raise IOError("IOC is not {}".format("started" if is_start else "stopped"))
+
+
+def is_ioc_up(ioc_name):
+    """
+    Determine if IOC is up by checking for the existence of its heartbeat PV
+    Args:
+        ioc_name: IOC to check
+
+    Returns: True if IOC is up; False otherwise
+    """
+    return g.get_pv("AS:{}:SR_heartbeat".format(ioc_name), is_local=True) is not None
